fix: stop a commented-out style tag from silencing the rest of a book
strip_tags had no comment state and cleared its tag buffer on every '<', so the
'<style>' inside an ordinary HTML comment opened a suppression that the '>' of
'-->' could not close. Every word after it stopped counting — and because the
estimate is persisted on books.word_count, the truncated figure stuck until the
file's mtime or size moved.

The migration's scan_roots guard was a tautology: books.library_id is NOT NULL
with an enforced foreign key, so an EXISTS on the id alone excludes nothing.
The backfill's work set is the *configured* ebook_library_path, and
prune_orphan_libraries deliberately keeps a root that still owns books after
the library is repointed away from it — those rows were nulled with nothing
able to refill them, which is the exact hazard rule 06 warns about.

The backfill tests move to a sibling tests.rs, per rule 03 and to bring the
file back under rule 05's size cap.
